NDAX: move supporting classes to ExchangeNDAXAPI class
- additional capitalization fixes
