+++ ExchangeSharp/API/Exchanges/NDAX/ExchangeNDAXAPI.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using ExchangeSharp.NDAX;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
