+++ ExchangeSharp/API/Exchanges/NDAX/Models/AuthenticateResult.cs
@@ -1,11 +1,14 @@
-namespace ExchangeSharp.NDAX
+namespace ExchangeSharp
 {
-    public class AuthenticateResult
-    {
-        public bool Authenticated { get; set; }
-        public int UserId { get; set; }
-        public string Token { get; set; }
-        public string AccountId { get; set; }
-        public string OMSId { get; set; }
-    }
+	public sealed partial class ExchangeNDAXAPI
+	{
+		class AuthenticateResult
+		{
+			public bool Authenticated { get; set; }
+			public int UserId { get; set; }
+			public string Token { get; set; }
+			public string AccountId { get; set; }
+			public string OMSId { get; set; }
+		}
+	}
 }
