+++ ExchangeSharp/API/Exchanges/NDAX/Models/GenericResponse.cs
@@ -1,16 +1,19 @@
 using Newtonsoft.Json;
 
-namespace ExchangeSharp.NDAX
+namespace ExchangeSharp
 {
-    public class GenericResponse
-    {
-        [JsonProperty("result")]
-        public bool Result { get; set; }
-        [JsonProperty("errormsg")]
-        public string ErrorMsg { get; set; }
-        [JsonProperty("errorcode")]
-        public int ErrorCode { get; set; }
-        [JsonProperty("detail")]
-        public int Detail { get; set; }
-    }
+	public sealed partial class ExchangeNDAXAPI
+	{
+		class GenericResponse
+		{
+			[JsonProperty("result")]
+			public bool Result { get; set; }
+			[JsonProperty("errormsg")]
+			public string ErrorMsg { get; set; }
+			[JsonProperty("errorcode")]
+			public int ErrorCode { get; set; }
+			[JsonProperty("detail")]
+			public int Detail { get; set; }
+		}
+	}
 }
