+++ ExchangeSharp/API/Exchanges/NDAX/Models/MessageType.cs
@@ -1,35 +1,38 @@
-namespace ExchangeSharp.NDAX
+namespace ExchangeSharp
 {
-    public enum MessageType
-    {
-        /// <summary>
-        /// Request.
-        /// </summary>
-        Request = 0,
+	public sealed partial class ExchangeNDAXAPI
+	{
+		enum MessageType
+		{
+			/// <summary>
+			/// Request.
+			/// </summary>
+			Request = 0,
 
-        /// <summary>
-        /// Reply.
-        /// </summary>
-        Reply = 1,
+			/// <summary>
+			/// Reply.
+			/// </summary>
+			Reply = 1,
 
-        /// <summary>
-        /// Subscribe to event.
-        /// </summary>
-        SubscribeToEvent = 2,
+			/// <summary>
+			/// Subscribe to event.
+			/// </summary>
+			SubscribeToEvent = 2,
 
-        /// <summary>
-        /// Event.
-        /// </summary>
-        Event = 3,
+			/// <summary>
+			/// Event.
+			/// </summary>
+			Event = 3,
 
-        /// <summary>
-        /// Unsubscribe from event.
-        /// </summary>
-        UnsubscribeFromEvent = 4,
+			/// <summary>
+			/// Unsubscribe from event.
+			/// </summary>
+			UnsubscribeFromEvent = 4,
 
-        /// <summary>
-        /// Error.
-        /// </summary>
-        Error = 5
-    }
+			/// <summary>
+			/// Error.
+			/// </summary>
+			Error = 5
+		}
+	}
 }
