+++ ExchangeSharp/API/Exchanges/NDAX/Models/SendOrderResponse.cs
@@ -1,14 +1,17 @@
 using Newtonsoft.Json;
 
-namespace ExchangeSharp.NDAX
+namespace ExchangeSharp
 {
-    public class SendOrderResponse
-    {
-        [JsonProperty("errormsg")]
-        public string ErrorMsg { get; set; }
-        [JsonProperty("status")]
-        public string Status { get; set; }
-        [JsonProperty("OrderId")]
-        public int OrderId { get; set; }
-    }
+	public sealed partial class ExchangeNDAXAPI
+	{
+		class SendOrderResponse
+		{
+			[JsonProperty("errormsg")]
+			public string ErrorMsg { get; set; }
+			[JsonProperty("status")]
+			public string Status { get; set; }
+			[JsonProperty("OrderId")]
+			public int OrderId { get; set; }
+		}
+	}
 }
